Remove std::string from reader dependency
